API updates and additions
Planned API endpoints and created placeholder classes for their
controllers. Updated unit tests to test existing API code.

+++ test/api/BoardsTest.php
@@ -1,24 +1,41 @@
 <?php
-require 'Mocks.php';
+require_once 'Mocks.php';
 
 class BoardsTest extends PHPUnit_Framework_TestCase {
 
     public static function setupBeforeClass() {
-        RedBeanPHP\R::setup('sqlite:tests.db');
+        try {
+            RedBeanPHP\R::setup('sqlite:tests.db');
+        } catch (Exception $ex) { }
     }
 
     public static function tearDownAfterClass() {
-        unlink('tests.db');
+        if (file_exists('tests.db')) {
+            unlink('tests.db');
+        }
     }
 
     public function testGetAllBoards() {
-        $boards = new Boards(new ControllerMock());
+        $boards = new Boards(new ContainerMock());
 
         $expected = new ApiJson();
         $expected->addAlert('info', 'No boards in database.');
 
         $this->assertEquals($expected,
             $boards->getAllBoards(null, new ResponseMock(), null));
+    }
+
+    public function testGetBoard() {
+        $boards = new Boards(new ContainerMock());
+
+        $expected = new ApiJson();
+        $expected->addAlert('error', 'No board found for ID 1.');
+
+        $args = [];
+        $args['id'] = '1';
+
+        $this->assertEquals($expected,
+            $boards->getBoard(null, new ResponseMock(), $args));
     }
 
 }
